引入新版序列化器，结合自定义编码解码器，进行简单的客户端+主机Netty通信，作为后面RPC框架的模板

+++ src/main/java/netty/annotation/SocketModule.java
@@ -13,5 +13,5 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 public @interface SocketModule {
-    int moduleId();
+    short moduleId();
 }
